refactor(toolkit): extract shared SMTP mailbox sanitization utilities
Move stripHeaderControlChars, sanitizeMailboxDisplayName,
sanitizeMailboxAddress, formatMailbox, and parseSendFrom into
@logto/connector-kit so connector-smtp2go-email and connector-mailjunky
share one implementation for header-injection prevention.
