+++ packages/toolkit/connector-kit/src/index.ts
@@ -14,6 +14,7 @@ import {
 } from './types/index.js';
 
 export * from './types/index.js';
+export * from './smtp-mailbox.js';
 
 export function validateConfig<Output, Input = Output>(
   config: unknown,
